# [dotnet] [bidi] Avoid intermediate JsonDocument allocation to determine unordered discriminator