--- a/dotnet/src/webdriver/BiDi/Communication/Json/Converters/Polymorphic/RealmInfoConverter.cs
+++ b/dotnet/src/webdriver/BiDi/Communication/Json/Converters/Polymorphic/RealmInfoConverter.cs
@@ -17,6 +17,7 @@
 // under the License.
 // </copyright>
 
+using OpenQA.Selenium.BiDi.Communication.Json.Converters.Polymorphic.Internal;
 using OpenQA.Selenium.BiDi.Modules.Script;
 using System;
 using System.Text.Json;
@@ -29,18 +30,16 @@ internal class RealmInfoConverter : JsonConverter<RealmInfo>
 {
     public override RealmInfo? Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
     {
-        var jsonDocument = JsonDocument.ParseValue(ref reader);
-
-        return jsonDocument.RootElement.GetProperty("type").ToString() switch
+        return reader.GetDiscriminator("type") switch
         {
-            "window" => jsonDocument.Deserialize<WindowRealmInfo>(options),
-            "dedicated-worker" => jsonDocument.Deserialize<DedicatedWorkerRealmInfo>(options),
-            "shared-worker" => jsonDocument.Deserialize<SharedWorkerRealmInfo>(options),
-            "service-worker" => jsonDocument.Deserialize<ServiceWorkerRealmInfo>(options),
-            "worker" => jsonDocument.Deserialize<WorkerRealmInfo>(options),
-            "paint-worklet" => jsonDocument.Deserialize<PaintWorkletRealmInfo>(options),
-            "audio-worklet" => jsonDocument.Deserialize<AudioWorkletRealmInfo>(options),
-            "worklet" => jsonDocument.Deserialize<WorkletRealmInfo>(options),
+            "window" => JsonSerializer.Deserialize<WindowRealmInfo>(ref reader, options),
+            "dedicated-worker" => JsonSerializer.Deserialize<DedicatedWorkerRealmInfo>(ref reader, options),
+            "shared-worker" => JsonSerializer.Deserialize<SharedWorkerRealmInfo>(ref reader, options),
+            "service-worker" => JsonSerializer.Deserialize<ServiceWorkerRealmInfo>(ref reader, options),
+            "worker" => JsonSerializer.Deserialize<WorkerRealmInfo>(ref reader, options),
+            "paint-worklet" => JsonSerializer.Deserialize<PaintWorkletRealmInfo>(ref reader, options),
+            "audio-worklet" => JsonSerializer.Deserialize<AudioWorkletRealmInfo>(ref reader, options),
+            "worklet" => JsonSerializer.Deserialize<WorkletRealmInfo>(ref reader, options),
             _ => null,
         };
     }