--- a/dotnet/src/webdriver/BiDi/Communication/Json/Converters/Polymorphic/RemoteValueConverter.cs
+++ b/dotnet/src/webdriver/BiDi/Communication/Json/Converters/Polymorphic/RemoteValueConverter.cs
@@ -17,6 +17,7 @@
 // under the License.
 // </copyright>
 
+using OpenQA.Selenium.BiDi.Communication.Json.Converters.Polymorphic.Internal;
 using OpenQA.Selenium.BiDi.Modules.Script;
 using System;
 using System.Text.Json;
@@ -29,41 +30,39 @@ internal class RemoteValueConverter : JsonConverter<RemoteValue>
 {
     public override RemoteValue? Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
     {
-        var jsonDocument = JsonDocument.ParseValue(ref reader);
-
-        if (jsonDocument.RootElement.ValueKind == JsonValueKind.String)
+        if (reader.TokenType == JsonTokenType.String)
         {
-            return new StringRemoteValue(jsonDocument.RootElement.GetString()!);
+            return new StringRemoteValue(reader.GetString()!);
         }
 
-        return jsonDocument.RootElement.GetProperty("type").ToString() switch
+        return reader.GetDiscriminator("type") switch
         {
-            "number" => jsonDocument.Deserialize<NumberRemoteValue>(options),
-            "boolean" => jsonDocument.Deserialize<BooleanRemoteValue>(options),
-            "bigint" => jsonDocument.Deserialize<BigIntRemoteValue>(options),
-            "string" => jsonDocument.Deserialize<StringRemoteValue>(options),
-            "null" => jsonDocument.Deserialize<NullRemoteValue>(options),
-            "undefined" => jsonDocument.Deserialize<UndefinedRemoteValue>(options),
-            "symbol" => jsonDocument.Deserialize<SymbolRemoteValue>(options),
-            "array" => jsonDocument.Deserialize<ArrayRemoteValue>(options),
-            "object" => jsonDocument.Deserialize<ObjectRemoteValue>(options),
-            "function" => jsonDocument.Deserialize<FunctionRemoteValue>(options),
-            "regexp" => jsonDocument.Deserialize<RegExpRemoteValue>(options),
-            "date" => jsonDocument.Deserialize<DateRemoteValue>(options),
-            "map" => jsonDocument.Deserialize<MapRemoteValue>(options),
-            "set" => jsonDocument.Deserialize<SetRemoteValue>(options),
-            "weakmap" => jsonDocument.Deserialize<WeakMapRemoteValue>(options),
-            "weakset" => jsonDocument.Deserialize<WeakSetRemoteValue>(options),
-            "generator" => jsonDocument.Deserialize<GeneratorRemoteValue>(options),
-            "error" => jsonDocument.Deserialize<ErrorRemoteValue>(options),
-            "proxy" => jsonDocument.Deserialize<ProxyRemoteValue>(options),
-            "promise" => jsonDocument.Deserialize<PromiseRemoteValue>(options),
-            "typedarray" => jsonDocument.Deserialize<TypedArrayRemoteValue>(options),
-            "arraybuffer" => jsonDocument.Deserialize<ArrayBufferRemoteValue>(options),
-            "nodelist" => jsonDocument.Deserialize<NodeListRemoteValue>(options),
-            "htmlcollection" => jsonDocument.Deserialize<HtmlCollectionRemoteValue>(options),
-            "node" => jsonDocument.Deserialize<NodeRemoteValue>(options),
-            "window" => jsonDocument.Deserialize<WindowProxyRemoteValue>(options),
+            "number" => JsonSerializer.Deserialize<NumberRemoteValue>(ref reader, options),
+            "boolean" => JsonSerializer.Deserialize<BooleanRemoteValue>(ref reader, options),
+            "bigint" => JsonSerializer.Deserialize<BigIntRemoteValue>(ref reader, options),
+            "string" => JsonSerializer.Deserialize<StringRemoteValue>(ref reader, options),
+            "null" => JsonSerializer.Deserialize<NullRemoteValue>(ref reader, options),
+            "undefined" => JsonSerializer.Deserialize<UndefinedRemoteValue>(ref reader, options),
+            "symbol" => JsonSerializer.Deserialize<SymbolRemoteValue>(ref reader, options),
+            "array" => JsonSerializer.Deserialize<ArrayRemoteValue>(ref reader, options),
+            "object" => JsonSerializer.Deserialize<ObjectRemoteValue>(ref reader, options),
+            "function" => JsonSerializer.Deserialize<FunctionRemoteValue>(ref reader, options),
+            "regexp" => JsonSerializer.Deserialize<RegExpRemoteValue>(ref reader, options),
+            "date" => JsonSerializer.Deserialize<DateRemoteValue>(ref reader, options),
+            "map" => JsonSerializer.Deserialize<MapRemoteValue>(ref reader, options),
+            "set" => JsonSerializer.Deserialize<SetRemoteValue>(ref reader, options),
+            "weakmap" => JsonSerializer.Deserialize<WeakMapRemoteValue>(ref reader, options),
+            "weakset" => JsonSerializer.Deserialize<WeakSetRemoteValue>(ref reader, options),
+            "generator" => JsonSerializer.Deserialize<GeneratorRemoteValue>(ref reader, options),
+            "error" => JsonSerializer.Deserialize<ErrorRemoteValue>(ref reader, options),
+            "proxy" => JsonSerializer.Deserialize<ProxyRemoteValue>(ref reader, options),
+            "promise" => JsonSerializer.Deserialize<PromiseRemoteValue>(ref reader, options),
+            "typedarray" => JsonSerializer.Deserialize<TypedArrayRemoteValue>(ref reader, options),
+            "arraybuffer" => JsonSerializer.Deserialize<ArrayBufferRemoteValue>(ref reader, options),
+            "nodelist" => JsonSerializer.Deserialize<NodeListRemoteValue>(ref reader, options),
+            "htmlcollection" => JsonSerializer.Deserialize<HtmlCollectionRemoteValue>(ref reader, options),
+            "node" => JsonSerializer.Deserialize<NodeRemoteValue>(ref reader, options),
+            "window" => JsonSerializer.Deserialize<WindowProxyRemoteValue>(ref reader, options),
             _ => null,
         };
     }