### **User description**
### Motivation and Context
Improve memory allocation.

Before:
```
| Method               | Mean     | Error    | StdDev   | Gen0    | Gen1    | Allocated |
|--------------------- |---------:|---------:|---------:|--------:|--------:|----------:|
|     LocateNodesAsync | 12.18 ms | 0.060 ms | 0.036 ms | 78.1250 | 31.2500 | 532.29 KB |
```
After:
```
| Method               | Mean     | Error    | StdDev   | Gen0    | Gen1    | Allocated |
|--------------------- |---------:|---------:|---------:|--------:|--------:|----------:|
|     LocateNodesAsync | 12.12 ms | 0.062 ms | 0.037 ms | 62.5000 | 15.6250 | 404.28 KB |
```

### Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

### Checklist
<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] I have read the [contributing](https://github.com/SeleniumHQ/selenium/blob/trunk/CONTRIBUTING.md) document.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [ ] I have added tests to cover my changes.
- [x] All new and existing tests passed.
<!--- Provide a general summary of your changes in the Title above -->


___

### **PR Type**
Enhancement


___

### **Description**
- Replaced `JsonDocument` usage with `Utf8JsonReader` for better performance.

- Introduced `GetDiscriminator` extension method for cleaner discriminator extraction.

- Reduced memory allocation in JSON deserialization processes.

- Updated multiple converters to use the new `GetDiscriminator` method.


___



### **Changes walkthrough** 📝
<table><thead><tr><th></th><th align="left">Relevant files</th></tr></thead><tbody><tr><td><strong>Enhancement</strong></td><td><table>
<tr>
  <td>
    <details>
      <summary><strong>EvaluateResultConverter.cs</strong><dd><code>Optimize `EvaluateResult` deserialization with `Utf8JsonReader`</code></dd></summary>
<hr>

dotnet/src/webdriver/BiDi/Communication/Json/Converters/Polymorphic/EvaluateResultConverter.cs

<li>Replaced <code>JsonDocument</code> with <code>Utf8JsonReader</code> for deserialization.<br> <li> Utilized <code>GetDiscriminator</code> for discriminator extraction.<br> <li> Improved memory efficiency in <code>EvaluateResult</code> deserialization.


</details>


  </td>
  <td><a href="https://github.com/SeleniumHQ/selenium/pull/15555/files#diff-2b4700f9d83b3c890a0eac145398021e0ab430758d441156ab88f84ba40ff2cd">+4/-5</a>&nbsp; &nbsp; &nbsp; </td>

</tr>

<tr>
  <td>
    <details>
      <summary><strong>LogEntryConverter.cs</strong><dd><code>Enhance `LogEntry` deserialization with `Utf8JsonReader`</code>&nbsp; </dd></summary>
<hr>

dotnet/src/webdriver/BiDi/Communication/Json/Converters/Polymorphic/LogEntryConverter.cs

<li>Switched to <code>Utf8JsonReader</code> for <code>LogEntry</code> deserialization.<br> <li> Replaced <code>JsonDocument</code> with <code>GetDiscriminator</code> for type handling.


</details>


  </td>
  <td><a href="https://github.com/SeleniumHQ/selenium/pull/15555/files#diff-573d86ca0ef4e4549d1540bcb3d744b2e56e66c49daef7758ab5232ee50aa8c8">+4/-5</a>&nbsp; &nbsp; &nbsp; </td>

</tr>

<tr>
  <td>
    <details>
      <summary><strong>MessageConverter.cs</strong><dd><code>Improve `Message` deserialization using `Utf8JsonReader`</code>&nbsp; </dd></summary>
<hr>

dotnet/src/webdriver/BiDi/Communication/Json/Converters/Polymorphic/MessageConverter.cs

<li>Replaced <code>JsonDocument</code> with <code>Utf8JsonReader</code> for <code>Message</code> deserialization.<br> <li> Used <code>GetDiscriminator</code> for cleaner type extraction.


</details>


  </td>
  <td><a href="https://github.com/SeleniumHQ/selenium/pull/15555/files#diff-a4036363ebefc63aba327f06b0c9427697f86735556dba4ea2e98241b3ca856c">+5/-7</a>&nbsp; &nbsp; &nbsp; </td>

</tr>

<tr>
  <td>
    <details>
      <summary><strong>RealmInfoConverter.cs</strong><dd><code>Optimize `RealmInfo` deserialization with `Utf8JsonReader`</code></dd></summary>
<hr>

dotnet/src/webdriver/BiDi/Communication/Json/Converters/Polymorphic/RealmInfoConverter.cs

<li>Updated <code>RealmInfo</code> deserialization to use <code>Utf8JsonReader</code>.<br> <li> Integrated <code>GetDiscriminator</code> for type determination.


</details>


  </td>
  <td><a href="https://github.com/SeleniumHQ/selenium/pull/15555/files#diff-863f352cd7277ca3aa65ea95702e0eaa8bae0d3136c6e09d467f068bf45ebe71">+10/-11</a>&nbsp; </td>

</tr>

<tr>
  <td>
    <details>
      <summary><strong>RemoteValueConverter.cs</strong><dd><code>Enhance `RemoteValue` deserialization with `Utf8JsonReader`</code></dd></summary>
<hr>

dotnet/src/webdriver/BiDi/Communication/Json/Converters/Polymorphic/RemoteValueConverter.cs

<li>Replaced <code>JsonDocument</code> with <code>Utf8JsonReader</code> for <code>RemoteValue</code> <br>deserialization.<br> <li> Enhanced type handling using <code>GetDiscriminator</code>.<br> <li> Improved memory efficiency for string and type-based deserialization.


</details>


  </td>
  <td><a href="https://github.com/SeleniumHQ/selenium/pull/15555/files#diff-c39cd7a227ded42315d504f1a8fa4d7e6de0f8cf429be68bdd0d0331ad1590eb">+30/-31</a>&nbsp; </td>

</tr>

<tr>
  <td>
    <details>
      <summary><strong>JsonExtensions.cs</strong><dd><code>Introduce `GetDiscriminator` extension for `Utf8JsonReader`</code></dd></summary>
<hr>

dotnet/src/webdriver/BiDi/Communication/Json/Internal/JsonExtensions.cs

<li>Added <code>GetDiscriminator</code> extension method for <code>Utf8JsonReader</code>.<br> <li> Simplified discriminator extraction logic for JSON objects.


</details>


  </td>
  <td><a href="https://github.com/SeleniumHQ/selenium/pull/15555/files#diff-8c683fe394e8363ac3fca2499e368fcbac261a35dc023e4734ae11cfa49b7915">+53/-0</a>&nbsp; &nbsp; </td>

</tr>
</table></td></tr></tr></tbody></table>

___

> <details> <summary>  Need help?</summary><li>Type <code>/help how to ...</code> in the comments thread for any questions about Qodo Merge usage.</li><li>Check out the <a href="https://qodo-merge-docs.qodo.ai/usage-guide/">documentation</a> for more information.</li></details>